add pacman-git changes
