+++ deps_git.go
@@ -1,4 +1,4 @@
-//go:build !next
+//go:build next
 
 package alpm
 
@@ -19,8 +19,8 @@ func (l DBList) FindSatisfier(depstring string) (IPackage, error) {
 
 	defer C.free(unsafe.Pointer(cDepString))
 
-	pkgList := (*C.struct___alpm_list_t)(unsafe.Pointer(l.list))
-	pkgHandle := (*C.struct___alpm_handle_t)(unsafe.Pointer(l.handle.ptr))
+	pkgList := (*C.struct__alpm_list_t)(unsafe.Pointer(l.list))
+	pkgHandle := (*C.struct__alpm_handle_t)(unsafe.Pointer(l.handle.ptr))
 
 	ptr := C.alpm_find_dbs_satisfier(pkgHandle, pkgList, cDepString)
 	if ptr == nil {
@@ -37,7 +37,7 @@ func (l PackageList) FindSatisfier(depstring string) (IPackage, error) {
 
 	defer C.free(unsafe.Pointer(cDepString))
 
-	pkgList := (*C.struct___alpm_list_t)(unsafe.Pointer(l.list))
+	pkgList := (*C.struct__alpm_list_t)(unsafe.Pointer(l.list))
 
 	ptr := C.alpm_find_satisfier(pkgList, cDepString)
 	if ptr == nil {
